==> 0-LoadTheData.Rmd <==
---
title: "Task0-LoadTheData"
author: "Grace Pehl, PhD"
date: "October 10, 2015"
output: html_document
---

```{r Download, echo=FALSE}
if (!file.exists('yelp_academic_dataset_review.json')) {
    dataURL = 'https://d396qusza40orc.cloudfront.net/dsscapstone/dataset/yelp_dataset_challenge_academic_dataset.zip'
    filename = 'yelp_dataset_challenge_academic_dataset.zip'
    download.file(dataURL, filename, mode = "w")
    unzip(filename, overwrite = FALSE)
}
```

```{r ReadJSON, cache=TRUE}
library(R.utils)
reviewfile = 'yelp_academic_dataset_review.json'
n_reviews <- countLines(reviewfile) # 1,569,264 entries
library(readr)
# Hadley Wickham's readr package saved hours of loading time
t0 <- proc.time()
reviews <- read_lines(reviewfile, n_max = n_reviews) # char vector
proc.time() - t0 # only 21s!!!
reviews <- sample(reviews, size = 0.1 * n_reviews) # subset
# pseudo-JSON, each entry is JSON, but entries aren't linked
# separate entries by commas and surround file with brackets
reviews <- paste(reviews, collapse = ", ")
reviews <- paste('[', reviews, ']', sep = '')
library(jsonlite)
reviews <- fromJSON(reviews) # dataframe
proc.time() - t0
saveRDS(reviews, file = 'reviews1.rds') #save the dataframe
```

```{r ReadJSON2, cache=TRUE}
bizfile = 'yelp_academic_dataset_business.json'
n_biz <- countLines(bizfile)
t1 <- proc.time()
businesses <- read_lines(bizfile, n_max = n_biz)
proc.time() - t1
businesses <- paste(businesses, collapse = ", ")
businesses <- paste('[', businesses, ']', sep = '')
businesses <- fromJSON(businesses)
proc.time() - t1
saveRDS(businesses, file = 'business1.rds')
```

```{r ReadJSON3, cache=TRUE}
tipfile = 'yelp_academic_dataset_tip.json'
n_tip <- countLines(tipfile)
t1 <- proc.time()
tips <- read_lines(tipfile, n_max = n_tip)
proc.time() - t1
tips <- paste(tips, collapse = ", ")
tips <- paste('[', tips, ']', sep = '')
tips <- fromJSON(tips)
proc.time() - t1
saveRDS(tips, file = 'tips1.rds')
```

```{r ReadJSON4, cache=TRUE}
userfile = 'yelp_academic_dataset_user.json'
n_user <- countLines(userfile)
t1 <- proc.time()
users <- read_lines(userfile, n_max = n_user)
proc.time() - t1
users <- paste(users, collapse = ", ")
users <- paste('[', users, ']', sep = '')
users <- fromJSON(users)
proc.time() - t1
saveRDS(users, file = 'users1.rds')
```

```{r FileExploration, echo=FALSE, eval=FALSE}
# quiz questions
reviews <- readLines(reviewfile, 100) 
reviews[100] #I have been coming ... almost 20 years
table(reviews$stars)[5] / length(reviews$stars) # 37% 5-star reviews
table(businesses$attributes$`Wi-Fi`)[1] / sum(table(businesses$attributes$`Wi-Fi`)) # 40% with free Wi-Fi
readLines(tipfile, n = 1000)[1000] # terrible service
which(users$compliments$funny > 10000) #33317
users$name[33317] #Brian
fisher.test(xtabs(~ (users$fans > 1) + (users$compliments$funny > 1)))
```
